Some vault test refactor
