+++ radix-engine-tests/tests/blueprints/vault/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod non_existent_vault;
 pub mod vault;
+pub mod vault_in_structs;
